+++ lib/employee.js
@@ -1,0 +1,13 @@
+class Employee {
+  constructor(name, id, email) {
+    this.name = name;
+    this.id = id;
+    this.email = email;
+  }
+  getName() {}
+  getId() {}
+  getEmail() {}
+  getRole() {
+    // return 'Employee'
+  }
+}
