+++ lib/engineer.js
@@ -1,0 +1,10 @@
+class Engineer extends Employee {
+    constructor (github){
+        super();
+        this.github = github;
+    }
+    getGithub();
+    getRole(){
+        // return 'Engineer'
+    };
+}
