+++ lib/intern.js
@@ -1,0 +1,9 @@
+class Intern extends Employee {
+  constructor(school) {
+    this.school = school;
+  }
+  getSchool() {}
+  getRole() {
+    // return 'Intern'
+  }
+}
