+++ lib/manager.js
@@ -1,0 +1,9 @@
+class Manager extends Employee {
+  constructor(officeNumber) {
+    super();
+    this.officeNumber = officeNumber;
+  }
+  getRole() {
+    // return 'Manager'
+  }
+}
