+++ src/index.js
@@ -1,1 +1,8 @@
+// imports
+const employee = require("../lib/employee");
+const engineer = require("../lib/engineer");
+const manager = require("../lib/manager");
+const intern = require("../lib/intern");
 
+// export
+module.exports = {};
